Remove hyphen from Single-File Component (#1785)
Compare the following:

https://vuejs.org/guide/scaling-up/sfc.html Single-File Components
https://vuejs.org/api/ Single File Component

This is a house style issue. A google search appears to show that this is a specific vuejs term, so the vuejs documentation should at the very least be consistent.
I would propose using the api form of "Single File Component" without the hyphen and so I have corrected this file accordingly. There are other places where this should also be corrected.